test_error_handling.py
"""Test error handling for Statista A2A agent.

This script tests that authentication errors and timeouts are properly
raised as exceptions and can be caught by the A2A framework.
"""

import asyncio
import logging
import os
from dotenv import load_dotenv

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Load environment
load_dotenv()


async def test_authentication_error():
    """Test that missing API key raises StatisaAuthenticationError."""
    print("\n=== Test 1: Missing API Key ===")

    # Save original key and remove it
    original_key = os.getenv("STATISTA_API_KEY")
    if "STATISTA_API_KEY" in os.environ:
        del os.environ["STATISTA_API_KEY"]

    try:
        from statista_agent.statista_tools import search_statistics, StatisaAuthenticationError
        from google.adk.tools.tool_context import ToolContext

        # Create a dummy tool context
        class DummyContext:
            def __init__(self):
                self.state = {}

        context = DummyContext()

        try:
            result = await search_statistics("test query", context, max_results=5)
            print(f"❌ FAILED: Should have raised StatisaAuthenticationError, got: {result}")
            return False
        except StatisaAuthenticationError as e:
            print(f"✅ PASSED: Correctly raised StatisaAuthenticationError: {e}")
            return True
        except Exception as e:
            print(f"❌ FAILED: Raised wrong exception type: {type(e).__name__}: {e}")
            return False
    finally:
        # Restore original key
        if original_key:
            os.environ["STATISTA_API_KEY"] = original_key


async def test_invalid_api_key():
    """Test that invalid API key raises StatisaAuthenticationError."""
    print("\n=== Test 2: Invalid API Key ===")

    # Save original key and set invalid one
    original_key = os.getenv("STATISTA_API_KEY")
    os.environ["STATISTA_API_KEY"] = "invalid_key_12345"

    # Clear the cached MCP client
    import statista_agent.statista_tools as tools
    tools._mcp_client = None

    try:
        from statista_agent.statista_tools import search_statistics, StatisaAuthenticationError
        from google.adk.tools.tool_context import ToolContext

        class DummyContext:
            def __init__(self):
                self.state = {}

        context = DummyContext()

        try:
            result = await search_statistics("test query", context, max_results=5)
            print(f"Result: {result[:200] if isinstance(result, str) else result}")
            print(f"⚠️  WARNING: Expected authentication error but got result. Check if API validates keys.")
            return True  # May be valid if API doesn't validate immediately
        except StatisaAuthenticationError as e:
            print(f"✅ PASSED: Correctly raised StatisaAuthenticationError: {e}")
            return True
        except Exception as e:
            # Check if it's an authentication-related error
            error_str = str(e).lower()
            if any(keyword in error_str for keyword in ['401', 'unauthorized', 'authentication', 'forbidden']):
                print(f"✅ PASSED: Raised authentication-related exception: {type(e).__name__}: {e}")
                return True
            print(f"❌ FAILED: Raised unexpected exception: {type(e).__name__}: {e}")
            return False
    finally:
        # Restore original key and clear cache
        if original_key:
            os.environ["STATISTA_API_KEY"] = original_key
        tools._mcp_client = None


async def test_valid_authentication():
    """Test that valid API key works correctly."""
    print("\n=== Test 3: Valid Authentication ===")

    api_key = os.getenv("STATISTA_API_KEY")
    if not api_key:
        print("⚠️  SKIPPED: No STATISTA_API_KEY set")
        return True

    # Clear the cached MCP client
    import statista_agent.statista_tools as tools
    tools._mcp_client = None

    try:
        from statista_agent.statista_tools import search_statistics, StatisaAuthenticationError

        class DummyContext:
            def __init__(self):
                self.state = {}

        context = DummyContext()

        try:
            result = await search_statistics("electric vehicles", context, max_results=3)
            if isinstance(result, str) and len(result) > 0:
                print(f"✅ PASSED: Successfully authenticated and got results")
                print(f"Result preview: {result[:200]}...")
                return True
            else:
                print(f"⚠️  WARNING: Got unexpected result type: {type(result)}")
                return True
        except StatisaAuthenticationError as e:
            print(f"❌ FAILED: Valid key raised authentication error: {e}")
            return False
        except Exception as e:
            print(f"❌ FAILED: Unexpected exception: {type(e).__name__}: {e}")
            return False
    finally:
        tools._mcp_client = None


async def main():
    """Run all error handling tests."""
    print("=" * 60)
    print("Testing Statista A2A Agent Error Handling")
    print("=" * 60)

    results = []

    # Test 1: Missing API key
    results.append(await test_authentication_error())

    # Test 2: Invalid API key
    results.append(await test_invalid_api_key())

    # Test 3: Valid authentication (if key is set)
    results.append(await test_valid_authentication())

    # Summary
    print("\n" + "=" * 60)
    print(f"Test Summary: {sum(results)}/{len(results)} tests passed")
    print("=" * 60)

    if all(results):
        print("✅ All tests passed!")
        print("\nYour error handling is now A2A-compliant:")
        print("  - Authentication errors are properly raised")
        print("  - The A2A framework will convert them to JSON-RPC errors")
        print("  - Clients will receive proper error codes and messages")
    else:
        print("❌ Some tests failed. Please review the errors above.")

    return all(results)


if __name__ == "__main__":
    success = asyncio.run(main())
    exit(0 if success else 1)
